Minor changes and refacto
